requirements.txt
azure-functions
requests
tabulate
xirr
